Split review GUI thumbnail_render into asset_usd and thumbnail_capture.
Separate USD path/AABB resolution from Kit viewport capture and wire
SimApp server and visualization_service to the new modules.
