+++ isaaclab_arena_examples/agentic_environment_generation/review_gui/simapp/asset_usd.py
@@ -1,0 +1,97 @@
+# Copyright (c) 2026, The Isaac Lab Arena Project Developers (https://github.com/isaac-sim/IsaacLab-Arena/blob/main/CONTRIBUTORS.md).
+# All rights reserved.
+#
+# SPDX-License-Identifier: Apache-2.0
+
+"""Resolve graph nodes to USD paths and local AABB dimensions (no Kit viewport)."""
+
+from __future__ import annotations
+
+import hashlib
+import sys
+
+from isaaclab_arena.environments.arena_env_graph_spec import ArenaEnvInitialGraphSpec
+from isaaclab_arena.environments.arena_env_graph_types import ArenaEnvGraphNodeType
+
+# Registry-backed nodes with a root USD. ``object_reference`` nodes point at a prim
+# inside a parent background and need parent-stage framing — not supported yet.
+RENDERABLE_NODE_TYPES = frozenset({
+    ArenaEnvGraphNodeType.EMBODIMENT,
+    ArenaEnvGraphNodeType.BACKGROUND,
+    ArenaEnvGraphNodeType.OBJECT,
+})
+
+AabbDimensionsM = tuple[float, float, float]
+
+
+def usd_cache_key(usd_path: str) -> str:
+    """Return a stable short hash for caching PNGs keyed by USD path."""
+    return hashlib.sha1(usd_path.encode("utf-8")).hexdigest()[:16]
+
+
+def resolve_node_usd_paths(spec: ArenaEnvInitialGraphSpec) -> dict[str, str]:
+    """Map ``node.id → usd_path`` via :class:`AssetRegistry`."""
+    try:
+        from isaaclab_arena.assets.registries import AssetRegistry  # noqa: PLC0415
+    except Exception as exc:
+        print(f"[asset_usd] AssetRegistry import failed: {exc}", file=sys.stderr)
+        return {}
+
+    registry = AssetRegistry()
+    paths: dict[str, str] = {}
+    for node in spec.nodes:
+        if node.type not in RENDERABLE_NODE_TYPES:
+            continue
+        try:
+            if not registry.is_registered(node.name):
+                print(f"[asset_usd]   {node.id}: asset '{node.name}' not registered, skipping.", file=sys.stderr)
+                continue
+            cls = registry.get_asset_by_name(node.name)
+            usd_path = extract_usd_path(cls)
+            if not usd_path:
+                print(f"[asset_usd]   {node.id}: '{node.name}' has no usd_path, skipping.", file=sys.stderr)
+                continue
+            paths[node.id] = usd_path
+        except Exception as exc:
+            print(f"[asset_usd]   {node.id}: lookup failed for '{node.name}': {exc}", file=sys.stderr)
+    return paths
+
+
+def extract_usd_path(cls) -> str | None:
+    """Return the asset's root USD path, or ``None`` if not extractable."""
+    usd_path = getattr(cls, "usd_path", None)
+    if usd_path:
+        return usd_path
+
+    try:
+        instance = cls()
+    except Exception:
+        return None
+    scene_config = getattr(instance, "scene_config", None)
+    robot = getattr(scene_config, "robot", None) if scene_config is not None else None
+    spawn = getattr(robot, "spawn", None) if robot is not None else None
+    return getattr(spawn, "usd_path", None) if spawn is not None else None
+
+
+def aabb_dimensions_from_usd(usd_path: str) -> AabbDimensionsM | None:
+    """Return local axis-aligned bounding box size (x, y, z) in meters for a USD asset."""
+    try:
+        from isaaclab_arena.utils.usd_helpers import compute_local_bounding_box_from_usd  # noqa: PLC0415
+
+        bbox = compute_local_bounding_box_from_usd(usd_path)
+        size = bbox.size[0]
+        return (float(size[0]), float(size[1]), float(size[2]))
+    except Exception as exc:
+        print(f"[asset_usd]   bbox failed for {usd_path}: {exc}", file=sys.stderr)
+        return None
+
+
+def resolve_node_aabb_dimensions_m(spec: ArenaEnvInitialGraphSpec) -> dict[str, AabbDimensionsM]:
+    """Return axis-aligned bounding box sizes in meters for each node with a resolvable USD."""
+    asset_paths = resolve_node_usd_paths(spec)
+    dimensions: dict[str, AabbDimensionsM] = {}
+    for node_id, usd_path in asset_paths.items():
+        dims = aabb_dimensions_from_usd(usd_path)
+        if dims is not None:
+            dimensions[node_id] = dims
+    return dimensions
